+++ streamlit_apps/🌊 epochs.py
@@ -11,6 +11,7 @@ eeg_epoch_visualization_dashboard()
 
 # Footer section
 version = get_version_from_pyproject()
+
 footer_html = f"""
     <div style='position: fixed; bottom: 0; left: 0; padding: 10px;'>
         <span>Version: {version}</span>